+++ apps/alpha-test-app/src/views/FormDemoView.vue
@@ -1,11 +1,13 @@
 <script lang="ts" setup>
 import { OnyxAppLayout, OnyxHeadline, OnyxPageLayout } from "sit-onyx";
+import { ref } from "vue";
 import { useI18n } from "vue-i18n";
+import LanguageSelection from "../components/LanguageSelection.vue";
 import FormDemo, { type FormData } from "../components/form-demo/FormDemo.vue";
 
 const { t, locale } = useI18n();
 
-const validFormData = {
+const validFormData = ref<FormData>({
   defaultInput: "No Validation",
   requiredInput: "Is filled",
   minlengthInput: "Is long enough",
@@ -13,17 +15,14 @@ const validFormData = {
   patternInput: "only lowercase or space",
   switch: true,
   checkboxGroup: [2],
-  radioGroup: { id: 1, label: "Option 1" },
-} satisfies FormData;
+  radioGroup: 1,
+});
 
-const invalidFormData = {
+const invalidFormData = ref<FormData>({
   defaultInput: "No Validation",
-  requiredInput: "",
-  minlengthInput: "",
   typeInput: "NotAmail",
   patternInput: "NO UPPERCASE ALLOWED",
-  switch: false,
-} satisfies FormData;
+});
 </script>
 
 <template>
@@ -39,12 +38,14 @@ const invalidFormData = {
 
       <div class="page">
         <OnyxHeadline is="h1" element="h1">Initially Invalid example</OnyxHeadline>
-        <FormDemo :form-data="invalidFormData" />
+        <FormDemo v-model="invalidFormData" />
+        <pre class="state">State: {{ invalidFormData }}</pre>
 
         <hr />
 
         <OnyxHeadline is="h1" element="h1">Initially Valid example</OnyxHeadline>
-        <FormDemo :form-data="validFormData" />
+        <FormDemo v-model="validFormData" />
+        <pre class="state">State: {{ validFormData }}</pre>
       </div>
     </OnyxPageLayout>
   </OnyxAppLayout>
@@ -56,5 +57,9 @@ const invalidFormData = {
 }
 .sidebar {
   padding: var(--onyx-spacing-3xs);
+}
+.state {
+  color: var(--onyx-color-text-icons-neutral-soft);
+  padding: var(--onyx-spacing-md) 0;
 }
 </style>