+++ packages/sit-onyx/src/components/OnyxListbox/OnyxListbox.vue
@@ -1,13 +1,14 @@
 <script
   lang="ts"
   setup
-  generic="TValue extends ListboxValue = ListboxValue, TMultiple extends boolean = false"
+  generic="TValue extends SelectOptionValue = SelectOptionValue, TMultiple extends boolean = false"
 >
-import { createId, createListbox, type ListboxValue } from "@sit-onyx/headless";
+import { createId, createListbox } from "@sit-onyx/headless";
 import { computed, ref, watch, watchEffect } from "vue";
 import { useCheckAll } from "../../composables/checkAll";
 import { useScrollEnd } from "../../composables/scrollEnd";
 import { injectI18n } from "../../i18n";
+import type { SelectOptionValue } from "../../types";
 import OnyxEmpty from "../OnyxEmpty/OnyxEmpty.vue";
 import OnyxListboxOption from "../OnyxListboxOption/OnyxListboxOption.vue";
 import OnyxLoadingIndicator from "../OnyxLoadingIndicator/OnyxLoadingIndicator.vue";
@@ -123,7 +124,7 @@ const {
       return i.label.toLowerCase().trim().startsWith(label.toLowerCase());
     });
     if (!firstMatch) return;
-    activeOption.value = firstMatch.id;
+    activeOption.value = firstMatch.value;
   },
 });
 
@@ -145,7 +146,7 @@ const { vScrollEnd, isScrollEnd } = useScrollEnd({
 const isEmpty = computed(() => props.options.length === 0);
 
 const enabledOptionValues = computed(() =>
-  props.options.filter((i) => !i.disabled).map(({ id }) => id),
+  props.options.filter((i) => !i.disabled).map(({ value }) => value),
 );
 
 /**
@@ -227,19 +228,19 @@ watchEffect(() => {
 
         <OnyxListboxOption
           v-for="option in options"
-          :key="option.id.toString()"
+          :key="option.value.toString()"
           v-bind="
             headlessOption({
-              value: option.id,
+              value: option.value,
               label: option.label,
               disabled: option.disabled,
               selected:
-                option.id === props.modelValue ||
-                (Array.isArray(props.modelValue) && props.modelValue.includes(option.id)),
+                option.value === props.modelValue ||
+                (Array.isArray(props.modelValue) && props.modelValue.includes(option.value)),
             })
           "
           :multiple="props.multiple"
-          :active="option.id === activeOption"
+          :active="option.value === activeOption"
           :icon="option.icon"
           :color="option.color"
         >
@@ -247,13 +248,13 @@ watchEffect(() => {
         </OnyxListboxOption>
       </ul>
 
-      <li v-if="props.lazyLoading?.loading" class="onyx-listbox__slot">
+      <div v-if="props.lazyLoading?.loading" class="onyx-listbox__slot">
         <OnyxLoadingIndicator class="onyx-listbox__loading" />
-      </li>
+      </div>
 
-      <li v-if="slots.optionsEnd" class="onyx-listbox__slot">
+      <div v-if="slots.optionsEnd" class="onyx-listbox__slot">
         <slot name="optionsEnd"></slot>
-      </li>
+      </div>
     </div>
 
     <slot name="footer">