+++ packages/sit-onyx/src/components/OnyxListbox/types.ts
@@ -1,8 +1,13 @@
-import type { ListboxModelValue, ListboxValue } from "@sit-onyx/headless";
+import type { SelectOption, SelectOptionValue } from "../../types";
 import type { OnyxListboxOptionProps } from "../OnyxListboxOption/types";
 
+export type ListboxModelValue<
+  TValue extends SelectOptionValue = SelectOptionValue,
+  TMultiple extends boolean = false,
+> = TMultiple extends true ? TValue[] : TValue;
+
 export type OnyxListboxProps<
-  TValue extends ListboxValue = ListboxValue,
+  TValue extends SelectOptionValue = SelectOptionValue,
   TMultiple extends boolean = false,
 > = {
   /**
@@ -18,7 +23,7 @@ export type OnyxListboxProps<
    */
   message?: string;
   /**
-   * Current value / selected options.
+   * Current value / selected option(s).
    */
   modelValue?: ListboxModelValue<TValue, TMultiple>;
   /**
@@ -52,27 +57,16 @@ export type OnyxListboxProps<
   lazyLoading?: ListboxLazyLoading;
 };
 
-export type ListboxOption<T extends ListboxValue = ListboxValue> = Pick<
-  OnyxListboxOptionProps,
-  "color" | "icon"
-> & {
-  /**
-   * Option ID / value to use when the option is selected.
-   */
-  id: T;
-  /**
-   * Label to show.
-   */
-  label: string;
-  /**
-   * Whether the option is disabled.
-   */
-  disabled?: boolean;
-  /**
-   * Optional group name. If set, all options will be grouped under that group name.
-   */
-  group?: string;
-};
+export type ListboxOption<TValue extends SelectOptionValue = SelectOptionValue> = Pick<
+  SelectOption<TValue>,
+  "value" | "label" | "disabled"
+> &
+  Pick<OnyxListboxOptionProps, "color" | "icon"> & {
+    /**
+     * Optional group name. If set, all options will be grouped under that group name.
+     */
+    group?: string;
+  };
 
 export type ListboxLazyLoading = {
   /**