+++ packages/sit-onyx/src/components/OnyxUserMenu/OnyxUserMenu.stories.ts
@@ -37,8 +37,8 @@ export const Default = {
     username: "Jane Doe",
     description: "Company Name",
     options: [
-      { id: "/settings", label: "Settings", icon: settings },
-      { id: "logout", label: "Logout", icon: logout, color: "danger" },
+      { value: "/settings", label: "Settings", icon: settings },
+      { value: "logout", label: "Logout", icon: logout, color: "danger" },
     ],
     footer: () => h(() => ["App version", h("span", { class: "onyx-text--monospace" }, "1.0.0")]),
   },