+++ apps/playground/CHANGELOG.md
@@ -1,5 +1,33 @@
 # playground
 
+## 0.0.1-alpha.29
+
+### Patch Changes
+
+- Updated dependencies [57d81c9]
+  - sit-onyx@1.0.0-alpha.90
+
+## 0.0.1-alpha.28
+
+### Patch Changes
+
+- Updated dependencies [7951251]
+  - sit-onyx@1.0.0-alpha.89
+
+## 0.0.1-alpha.27
+
+### Patch Changes
+
+- Updated dependencies [8d65dce]
+  - sit-onyx@1.0.0-alpha.88
+
+## 0.0.1-alpha.26
+
+### Patch Changes
+
+- Updated dependencies [413d3e0]
+  - sit-onyx@1.0.0-alpha.87
+
 ## 0.0.1-alpha.25
 
 ### Patch Changes