+++ packages/chartjs-plugin/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @sit-onyx/chartjs-plugin
 
+## 1.0.0-alpha.45
+
+### Patch Changes
+
+- Updated dependencies [57d81c9]
+  - sit-onyx@1.0.0-alpha.90
+
+## 1.0.0-alpha.44
+
+### Patch Changes
+
+- Updated dependencies [7951251]
+  - sit-onyx@1.0.0-alpha.89
+
+## 1.0.0-alpha.43
+
+### Patch Changes
+
+- Updated dependencies [8d65dce]
+  - sit-onyx@1.0.0-alpha.88
+
+## 1.0.0-alpha.42
+
+### Patch Changes
+
+- Updated dependencies [413d3e0]
+  - sit-onyx@1.0.0-alpha.87
+
 ## 1.0.0-alpha.41
 
 ### Patch Changes