+++ packages/sit-onyx/CHANGELOG.md
@@ -1,5 +1,37 @@
 # sit-onyx
 
+## 1.0.0-alpha.90
+
+### Minor Changes
+
+- 57d81c9: feat: export `OnyxCheckbox` and `OnyxRadioButton`
+
+## 1.0.0-alpha.89
+
+### Major Changes
+
+- 7951251: refactor: align properties for selection components
+
+  - remove type `SelectionOption` and `ListboxOption` in favor of `SelectOption`. Affects: OnyxCheckboxGroup, OnyxRadioButtonGroup, OnyxListbox and OnyxSelect
+  - OnyxSelect: change property `modelValue` to be a select options instead of value only. This allows (e.g. to display a label for the current value(s) when the value(s) are not included in the options or the options are loading asynchronously)
+  - OnyxSelect: require property `options`
+  - OnyxRadioButtonGroup: change property `modelValue` to be the primitive value instead of a whole option
+  - fix(OnyxRadioButtonGroup): modelValue not shown as selected if type is number or boolean
+  - rename type `Multiple` to `SelectMultiple`
+  - remove utility type `TargetEvent`
+
+## 1.0.0-alpha.88
+
+### Minor Changes
+
+- 8d65dce: feat(OnyxTable): add row and column hover styles
+
+## 1.0.0-alpha.87
+
+### Minor Changes
+
+- 413d3e0: Implement external icon when external links are used inside OnyxNavItem
+
 ## 1.0.0-alpha.86
 
 ### Minor Changes